requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-generativeai==0.8.3
langchain==0.3.13
langchain-community==0.3.13
langchain-google-genai==2.0.8
chromadb==0.5.23
sentence-transformers==3.3.1
beautifulsoup4==4.12.3
requests==2.32.5
python-dotenv==1.1.1
Flask==3.1.0
flask-cors==5.0.0
stripe==13.0.1
anthropic==0.69.0
bcrypt==5.0.0
pyjwt==2.10.1
python-multipart==0.0.20
aiofiles==24.1.0